requirements.txt
django==3.2.25
djangorestframework==3.12.4
pandas==1.1.5
openpyxl==3.0.9
markdown==3.3.4
django-filter==2.4.0
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==5.3.0